requirements.txt
Django==3.1.5
django-jsonfield==1.4.1
djangorestframework==3.12.2
plaid-python==5.0.0
amqp==5.0.3
celery==5.0.5
SQLAlchemy==1.3.22
uuid==1.30